route-service/requirements.txt
fastapi
uvicorn[standard]
motor
odmantic
python-dotenv
pyjwt
prometheus-client
aio-pika